# Native engine is a stripped cdylib