@illicitonion 2018-03-07
LTO is indeed link time optimisation :)

(See https://github.com/pantsbuild/pants/pull/5560 for the PR which enables much more aggressive LTO, which we're probably not actually going to merge)